Use Cucumber and Aruba for acceptance tests.
It is not possible to create a testing suite using xcodeproj or an equivalent, or at least it was not obvious as to how to do so, so I've opted for copying a template Xcode project from features/fixtures to features/tmp, which I then operate on.
Use xcodebuild to compile pods and run tests.