Major refactoring. See full commit message.
Removed visitor pattern. Instead propagating runs through the test hierarchy, more akin to the implementation of Rspec. A reporter is passed from node to node when forwarding the run message. KISS.
A reporter is notified of events that occur during a run.
Modernized the project structure to include less folders, and to clarify the role of a reporter.
Updated unit tests to reflect changes.